Updated RBAC

+++ app/Http/Controllers/Blade/PermissionController.php
@@ -16,13 +16,19 @@ class PermissionController extends Controller
     // list of permissions
     public function index()
     {
-        $permissions = Permission::all();
+        if (!auth()->user()->can('permission.show'))
+            return abort(404);
+
+        $permissions = Permission::with('roles')->get();
         return view('pages.permissions.index',compact('permissions'));
     }
 
     // add permission page
     public function add()
     {
+        if (!auth()->user()->can('permission.add'))
+            return abort(404);
+
         return view('pages.permissions.add');
     }
 
@@ -44,8 +50,10 @@ class PermissionController extends Controller
     // edit page
     public function edit($id)
     {
-        $permission = Permission::findById($id);
+        if (!auth()->user()->can('permission.edit'))
+            return abort(404);
 
+        $permission = Permission::findById($id);
         return view('pages.permissions.edit',compact('permission'));
     }
 
@@ -70,6 +78,9 @@ class PermissionController extends Controller
     // delete permission
     public function destroy($id)
     {
+        if (!auth()->user()->can('permission.delete'))
+            return abort(404);
+
         $permission = Permission::findById($id);
         $permission->delete();
         return redirect()->back();
